+++ podcasts/Onboarding/Login/LoginLandingView.swift
@@ -331,7 +331,7 @@ private struct LoginButtons: View {
     }
 }
 
-private struct SocialLoginButtons: View {
+struct SocialLoginButtons: View {
     @EnvironmentObject var theme: Theme
     let coordinator: LoginCoordinator
 
@@ -341,7 +341,7 @@ private struct SocialLoginButtons: View {
             case .apple:
                 Button(L10n.socialSignInContinueWithApple) {
                     coordinator.signIn(with: provider)
-                }.buttonStyle(SocialButtonStyle(imageName: AppTheme.socialIconAppleImageName()))
+                }.buttonStyle(SocialButtonStyle(imageName: AppTheme.socialIconAppleImageName(theme: theme)))
             case .google:
                 Button(L10n.socialSignInContinueWithGoogle) {
                     coordinator.signIn(with: provider)