+++ podcasts/StarredFilterOverlayController.swift
@@ -62,6 +62,9 @@ class StarredFilterOverlayController: PCViewController {
     }
 
     private func setupNavBar() {
+        let backgroundColor = AppTheme.viewBackgroundColor()
+        changeNavTint(titleColor: AppTheme.colorForStyle(.primaryText01), iconsColor: AppTheme.colorForStyle(.primaryIcon03), backgroundColor: backgroundColor)
+
         largeTitleFont = UIFont.systemFont(ofSize: 22, weight: .bold)
 
         title = L10n.statusStarred